refactor(auth): clean up verification page logic
